refactor to prevent circular import
